fix imports and deprecate setDefaultFeer
