+++ packages/dappmanager/src/modules/stakerConfig/get/index.ts
@@ -1,0 +1,1 @@
+export { getStakerConfig } from "./getStakerConfig.js";
